+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Removed
+- `chafa-libload` feature, has been removed from ratatui-image. Simply use halfblocks directly.
+
+## [0.17.4] - 2025-12-25
+
+### Fixed
+- When entering search (both Link and slash), jump to the first match with the current scroll offset
+
 ## [0.17.3] - 2025-12-22
 
 ### Fixed